update overlays to use with routes
