+++ src/overlay/index.ts
@@ -1,9 +1,20 @@
-import _ from 'lodash';
 import Vue from 'vue';
-import Overlays from './index.vue';
 
 import { isAvailableVariable } from '../panel/helpers/isAvailableVariable';
 import translate from '../panel/helpers/translate';
+import VueRouter from 'vue-router';
+
+
+import moment from 'moment';
+import momentTimezone from 'moment-timezone';
+import VueMoment from 'vue-moment';
+import urlParam from '../panel/helpers/urlParam';
+
+Vue.use(VueMoment, {
+  moment, momentTimezone,
+});
+Vue.use(VueRouter);
+
 
 export interface Global {
   translations: any;
@@ -18,7 +29,7 @@ declare module 'vue/types/vue' {
     token: string;
     configuration: any;
     $moment?: any;
-    _: _.LoDashStatic;
+    urlParam(key: string): string | null;
   }
 }
 
@@ -29,14 +40,41 @@ const overlays = async () => {
   ]);
 
   Vue.prototype.translate = (v) => translate(v);
+  Vue.prototype.urlParam = (v) => urlParam(v);
+  Vue.prototype.token = token;
   Vue.prototype.token = token;
   Vue.prototype.configuration = global.configuration;
-  Vue.prototype._ = _;
 
-  new Vue({ // eslint-disable-line no-new
-    render(createElement) {
-      return createElement(Overlays, { props: { token } });
+  const router = new VueRouter({
+    mode: 'history',
+    base: __dirname,
+    routes: [
+      { path: '/overlays/alerts', name: 'alertsOverlay', component: () => import('./views/alerts.vue') },
+      { path: '/overlays/bets', name: 'betsOverlay', component: () => import('./views/bets.vue') },
+      { path: '/overlays/carousel', name: 'carouselOverlay', component: () => import('./views/carousel.vue') },
+      { path: '/overlays/clips', name: 'clipsOverlay', component: () => import('./views/clips.vue') },
+      { path: '/overlays/clipscarousel', name: 'clipsCarouselOverlay', component: () => import('./views/clipscarousel.vue') },
+      { path: '/overlays/credits', name: 'creditsOverlay', component: () => import('./views/credits.vue') },
+      { path: '/overlays/emotes', name: 'emotesOverlay', component: () => import('./views/emotes.vue') },
+      { path: '/overlays/eventlist', name: 'eventlistOverlay', component: () => import('./views/eventlist.vue') },
+      { path: '/overlays/goals', name: 'goalsOverlay', component: () => import('./views/goals.vue') },
+      { path: '/overlays/polls', name: 'pollsOverlay', component: () => import('./views/polls.vue') },
+      { path: '/overlays/stats', name: 'statsOverlay', component: () => import('./views/stats.vue') },
+      { path: '/overlays/text', name: 'textOverlay', component: () => import('./views/text.vue') },
+      { path: '/overlays/wheeloffortune', name: 'wheeloffortuneOverlay', component: () => import('./views/wheeloffortune.vue') },
+    ],
+  });
+
+  new Vue({
+    router,
+    created() {
+      this.$moment.locale(global.configuration.lang); // set proper moment locale
     },
+    template: `
+      <div id="app">
+        <router-view class="view"></router-view>
+      </div>
+    `,
   }).$mount('#overlays');
 };
