+++ src/overlay/views/goals.vue
@@ -133,10 +133,9 @@
 </template>
 
 <script lang="ts">
+import { Vue, Component } from 'vue-property-decorator';
 import io from 'socket.io-client';
 import safeEval from 'safe-eval'
-
-import Vue from 'vue';
 
 import BootstrapVue from 'bootstrap-vue'
 import 'bootstrap/dist/css/bootstrap.css'
@@ -154,40 +153,26 @@ Vue.use(VueMoment, {
 
 import { TweenLite } from 'gsap/TweenMax'
 
-export default Vue.extend({
-  props: ['token'],
-  data: function () {
-    const object: {
-      show: number,
-      group: Goals.Group | null,
-      goals: Goals.Goal[],
-      loadedFonts: string[]
-      socket: any,
-      lastSwapTime: number,
-      triggerUpdate: string[],
-      cssLoaded: string[],
-      current: { subscribers: number, followers: number },
-      interval: number[]
-    } = {
-      show: -1,
-      group: null,
-      goals: [],
-      socket: io('/overlays/goals', {query: "token=" + this.token}),
-      lastSwapTime: Date.now(),
-      loadedFonts: [],
-      triggerUpdate: [],
-      cssLoaded: [],
-      current: { subscribers: 0, followers: 0 },
-      interval: [],
-    };
-    return object
-  },
-  beforeDestroy: function() {
+@Component({})
+export default class GoalsOverlay extends Vue {
+  show: number = -1;
+  group: Goals.Group | null = null;
+  goals: Goals.Goal[] = [];
+  loadedFonts: string[] = [];
+  socket = io('/overlays/goals', {query: "token=" + this.token});
+  lastSwapTime: number = Date.now();
+  triggerUpdate: string[] = [];
+  cssLoaded: string[] = [];
+  current: { subscribers: number, followers: number } = { subscribers: 0, followers: 0 };
+  interval: number[] = [];
+
+  beforeDestroy() {
     for(const interval of this.interval) {
       clearInterval(interval);
     }
-  },
-  mounted: function () {
+  }
+
+  mounted() {
     this.$moment.locale(this.configuration.lang)
     this.refresh()
     this.interval.push(window.setInterval(() => this.refresh(), 5000));
@@ -204,169 +189,174 @@ export default Vue.extend({
         }
       }
     }, 100));
-  },
-  methods: {
-    beforeEnter: function (el) {
-      el.style.opacity = 0
-    },
-    doEnterAnimation: function (el, done) {
-      if (this.group === null) return
-      if (this.group.display.type === 'fade') {
-        TweenLite.to(el, (this.group.display.animationInMs || 1000) / 1000, {
-          opacity: 1,
-          onComplete: () => {
-            done()
-          }
-        })
-      }
-    },
-    doLeaveAnimation: function (el, done) {
-      if (this.group === null) return
-      if (this.group.display.type === 'fade') {
-        TweenLite.to(el, (this.group.display.animationOutMs || 1000) / 1000, {
-          opacity: 0,
-          onComplete: () => {
-            done()
-          }
-        })
-      }
-    },
-    isDisabled(idx) {
-      if (this.group === null) return false;
+  }
 
-      const goal = this.goals[idx]
-      return new Date(goal.endAfter).getTime() <= new Date().getTime() && !goal.endAfterIgnore
-    },
-    textStrokeGenerator(radius, color) {
-      if (radius === 0) return ''
+  beforeEnter (el) {
+    el.style.opacity = 0
+  }
 
-      // config
-      const steps = 30;
-      const blur = 2;
-      // generate text shadows, spread evenly around a circle
-      const radianStep = steps / (Math.PI * 2);
-      let cssStr = '';
-      for (let r=1; r <= radius; r++) {
-        for(let i=0; i < steps; i++) {
-          const curRads = radianStep * i;
-          const xOffset = (r * Math.sin(curRads)).toFixed(1);
-          const yOffset = (r * Math.cos(curRads)).toFixed(1);
-          if(i > 0 || r > 1) cssStr += ", ";
-          cssStr += xOffset + "px " + yOffset + "px " + blur + "px " + color;
+  doEnterAnimation (el, done) {
+    if (this.group === null) return
+    if (this.group.display.type === 'fade') {
+      TweenLite.to(el, (this.group.display.animationInMs || 1000) / 1000, {
+        opacity: 1,
+        onComplete: () => {
+          done()
         }
-      }
-      return cssStr
-    },
-    getFontFamilyCSS (family) {
-      return `"${family}" !important`;
-    },
-    refresh: function () {
-      const uid = window.location.href.split('/')[window.location.href.split('/').length - 1]
-      if (uid) {
-        this.socket.emit('current', (err, current: { subscribers: number, followers: number }) => {
-          if (err) return console.error(err)
-          this.current = current
-        })
-        this.socket.emit('findOne', { collection: 'groups', where: { uid }}, (err, cb: Goals.Group | null) => {
-          if (err) return console.error(err)
-          this.group = cb
-        })
-
-        this.socket.emit('find', { collection: 'goals', where: { groupId: uid }}, (err, goals: Goals.Goal[]) => {
-          if (err) return console.error(err)
-
-        // run check first
-        if (this.goals.length > 0) {
-          for (const goal of this.goals) {
-            let _goal = this._.find(this.goals, (o) => o.uid === goal.uid)
-            if (typeof _goal !== 'undefined') {
-              if (Number(_goal.currentAmount) !== Number(goal.currentAmount)) {
-                console.debug(_goal.currentAmount + ' => ' + goal.currentAmount)
-                this.triggerUpdate.push(goal.uid)
-              }
-            }
-          }
-        }
-
-        this.goals = goals
-
-        // update currentAmount for current types
-        for (const goal of this.goals) {
-          if (goal.type === 'currentFollowers') {
-            if (goal.currentAmount !== this.current.followers) this.triggerUpdate.push(goal.uid)
-            goal.currentAmount = this.current.followers
-          }
-          if (goal.type === 'currentSubscribers') {
-            if (goal.currentAmount !== this.current.subscribers) this.triggerUpdate.push(goal.uid)
-            goal.currentAmount = this.current.subscribers
-          }
-        }
-
-        // add css import
-        for (const goal of this.goals) {
-          if (!this.cssLoaded.includes(goal.uid)) {
-            this.cssLoaded.push(goal.uid);
-            const head = document.getElementsByTagName('head')[0]
-            const style = document.createElement('style')
-            style.type = 'text/css';
-            if (!this.loadedFonts.includes(goal.customization.css)) {
-              this.loadedFonts.push(goal.customization.css)
-              const css = goal.customization.css
-                .replace(/\#wrap/g, '#wrap-' + goal.uid) // replace .wrap with only this goal wrap
-              style.appendChild(document.createTextNode(css));
-            }
-            head.appendChild(style);
-          }
-        }
-
-        // add fonts import
-        const head = document.getElementsByTagName('head')[0]
-        const style = document.createElement('style')
-        style.type = 'text/css';
-
-        for (const goal of this.goals) {
-          if (!this.loadedFonts.includes(goal.customization.font.family)) {
-            this.loadedFonts.push(goal.customization.font.family)
-            const font = goal.customization.font.family.replace(/ /g, '+')
-            const css = "@import url('https://fonts.googleapis.com/css?family=" + font + "');"
-            style.appendChild(document.createTextNode(css));
-          }
-        }
-        head.appendChild(style);
-
-        // if custom html update all variables
-        for (const goal of this.goals) {
-          if (goal.display === 'custom') {
-            goal.customization.html = goal.customization.html
-              .replace(/\$name/g, goal.name)
-              .replace(/\$type/g, goal.type)
-              .replace(/\$goalAmount/g, String(goal.goalAmount))
-              .replace(/\$currentAmount/g, String(goal.currentAmount))
-              .replace(/\$percentageAmount/g, Number((100 / goal.goalAmount) * goal.currentAmount).toFixed())
-              .replace(/\$endAfter/g, goal.endAfter)
-          }
-
-          // trigger onUpdate on nextTick
-          this.$nextTick(() => {
-            if (this.triggerUpdate.includes(goal.uid)) {
-              const idx = this.triggerUpdate.indexOf(goal.uid);
-              this.triggerUpdate.splice(idx, 1);
-
-              console.debug('onUpdate : ' + goal.uid)
-              let toEval = `(function evaluation () { ${goal.customization.js}; onChange(${goal.currentAmount}) })()`
-              safeEval(toEval)
-            }
-          })
-        }
-
-          this.$nextTick(() => { if (this.show === -1) this.show = 0; })
-        })
-      } else {
-        console.error('Missing id param in url')
-      }
+      })
     }
   }
-})
+
+  doLeaveAnimation (el, done) {
+    if (this.group === null) return
+    if (this.group.display.type === 'fade') {
+      TweenLite.to(el, (this.group.display.animationOutMs || 1000) / 1000, {
+        opacity: 0,
+        onComplete: () => {
+          done()
+        }
+      })
+    }
+  }
+
+  isDisabled(idx) {
+    if (this.group === null) return false;
+
+    const goal = this.goals[idx]
+    return new Date(goal.endAfter).getTime() <= new Date().getTime() && !goal.endAfterIgnore
+  }
+
+  textStrokeGenerator(radius, color) {
+    if (radius === 0) return ''
+
+    // config
+    const steps = 30;
+    const blur = 2;
+    // generate text shadows, spread evenly around a circle
+    const radianStep = steps / (Math.PI * 2);
+    let cssStr = '';
+    for (let r=1; r <= radius; r++) {
+      for(let i=0; i < steps; i++) {
+        const curRads = radianStep * i;
+        const xOffset = (r * Math.sin(curRads)).toFixed(1);
+        const yOffset = (r * Math.cos(curRads)).toFixed(1);
+        if(i > 0 || r > 1) cssStr += ", ";
+        cssStr += xOffset + "px " + yOffset + "px " + blur + "px " + color;
+      }
+    }
+    return cssStr
+  }
+
+  getFontFamilyCSS (family) {
+    return `"${family}" !important`;
+  }
+
+  refresh () {
+    const uid = window.location.href.split('/')[window.location.href.split('/').length - 1]
+    if (uid) {
+      this.socket.emit('current', (err, current: { subscribers: number, followers: number }) => {
+        if (err) return console.error(err)
+        this.current = current
+      })
+      this.socket.emit('findOne', { collection: 'groups', where: { uid }}, (err, cb: Goals.Group | null) => {
+        if (err) return console.error(err)
+        this.group = cb
+      })
+
+      this.socket.emit('find', { collection: 'goals', where: { groupId: uid }}, (err, goals: Goals.Goal[]) => {
+        if (err) return console.error(err)
+
+      // run check first
+      if (this.goals.length > 0) {
+        for (const goal of this.goals) {
+          let _goal = this._.find(this.goals, (o) => o.uid === goal.uid)
+          if (typeof _goal !== 'undefined') {
+            if (Number(_goal.currentAmount) !== Number(goal.currentAmount)) {
+              console.debug(_goal.currentAmount + ' => ' + goal.currentAmount)
+              this.triggerUpdate.push(goal.uid)
+            }
+          }
+        }
+      }
+
+      this.goals = goals
+
+      // update currentAmount for current types
+      for (const goal of this.goals) {
+        if (goal.type === 'currentFollowers') {
+          if (goal.currentAmount !== this.current.followers) this.triggerUpdate.push(goal.uid)
+          goal.currentAmount = this.current.followers
+        }
+        if (goal.type === 'currentSubscribers') {
+          if (goal.currentAmount !== this.current.subscribers) this.triggerUpdate.push(goal.uid)
+          goal.currentAmount = this.current.subscribers
+        }
+      }
+
+      // add css import
+      for (const goal of this.goals) {
+        if (!this.cssLoaded.includes(goal.uid)) {
+          this.cssLoaded.push(goal.uid);
+          const head = document.getElementsByTagName('head')[0]
+          const style = document.createElement('style')
+          style.type = 'text/css';
+          if (!this.loadedFonts.includes(goal.customization.css)) {
+            this.loadedFonts.push(goal.customization.css)
+            const css = goal.customization.css
+              .replace(/\#wrap/g, '#wrap-' + goal.uid) // replace .wrap with only this goal wrap
+            style.appendChild(document.createTextNode(css));
+          }
+          head.appendChild(style);
+        }
+      }
+
+      // add fonts import
+      const head = document.getElementsByTagName('head')[0]
+      const style = document.createElement('style')
+      style.type = 'text/css';
+
+      for (const goal of this.goals) {
+        if (!this.loadedFonts.includes(goal.customization.font.family)) {
+          this.loadedFonts.push(goal.customization.font.family)
+          const font = goal.customization.font.family.replace(/ /g, '+')
+          const css = "@import url('https://fonts.googleapis.com/css?family=" + font + "');"
+          style.appendChild(document.createTextNode(css));
+        }
+      }
+      head.appendChild(style);
+
+      // if custom html update all variables
+      for (const goal of this.goals) {
+        if (goal.display === 'custom') {
+          goal.customization.html = goal.customization.html
+            .replace(/\$name/g, goal.name)
+            .replace(/\$type/g, goal.type)
+            .replace(/\$goalAmount/g, String(goal.goalAmount))
+            .replace(/\$currentAmount/g, String(goal.currentAmount))
+            .replace(/\$percentageAmount/g, Number((100 / goal.goalAmount) * goal.currentAmount).toFixed())
+            .replace(/\$endAfter/g, goal.endAfter)
+        }
+
+        // trigger onUpdate on nextTick
+        this.$nextTick(() => {
+          if (this.triggerUpdate.includes(goal.uid)) {
+            const idx = this.triggerUpdate.indexOf(goal.uid);
+            this.triggerUpdate.splice(idx, 1);
+
+            console.debug('onUpdate : ' + goal.uid)
+            let toEval = `(function evaluation () { ${goal.customization.js}; onChange(${goal.currentAmount}) })()`
+            safeEval(toEval)
+          }
+        })
+      }
+
+        this.$nextTick(() => { if (this.show === -1) this.show = 0; })
+      })
+    } else {
+      console.error('Missing id param in url')
+    }
+  }
+}
 </script>
 
 <style scoped>
